Remove unneeded lookup of piece color in Position.DoMove
When applying a move to the position, the color of the piece on
the from square should always be the color of the side to move.
And if the move is an attack, the color of the piece on the to
square is the opposing side. Using this realization, we can
eliminate looking up the color of the pieces on the from and to
squares, thus saving time.