Don't allocate memory for positions on the fly
Before, positions were being copied to the DoMove function, which then
made the move on the copied position, and returned a pointer to it.
This is unnecessarily slow since Go had to allocate memory for the
position argument of DoMove on the fly, and then copy the position state
over.

Instead, what is done now is a stack of positions are allocated before
the chess engine starts. During PERFT, this stack is indexed
by depth. When a move needs to be made at depth D, the current position's
state is copied to the position at index D on the stack, the move is made,
recursively explored, and then the old position state is copied from the
position object on the stack at depth D, back to the current position object.

This is much faster as now, memory for position object is not being allocated
on the fly, and runtime.mallocgc is called much less. And now only copying of
data is being done between position objects, instead of copying and allocation.

This dropped PERFT(6) from roughly ~8-7 seconds, to ~4 seconds.