Fix bugs in quiescence search and pv capturing
In quiescence search, checkmates detection logic was implemented,
which is not appropriate for quiescence search since only captures
and queen promotion moves are being generated.

at the start of each node, the pv for the node's ply needs to be
cleared. Not clearing this pv means that if we searched deeper in
the tree to the left of the current node, we might already have
moves being stored at the current ply, and higher plies. We don't
want to incorrectly copy these moves to the current pv we're
constructing, and we no longer need these moves. So clear them
from the pv at the current ply.